py_mcp_client/src/requirements.txt
fastmcp
pyyaml
prompt_toolkit